Add option to list available stable versions and URL
Option -available will send a HTTP GET request to the download page of
the official website of the Go project, then it will find all the HTML
tags identified by a CSS class name "toggle" or "toggleVisible" for the
latest stable version. Then it will process the text and print the list
in the console. Note that I am also modifying the content of the Git
ignore file to allow the addition of the test files, it seems that the
current exclusion rules are targetting everything inside the parent
source directory instead of the unnecessary files inside the child
source directory located inside src/gobu/