Modify available versions command to expect reader
To facilitate the testing of the functions used to extract the list of
stable versions from the download page of the official Go website this
commit will separate the HTTP request code from the code that parses and
extracts the relevant information from the HTML code. This way we can
pass a different buffer in the test files with static data and avoid
unnecessary requests to the live website in the continuous integration
service.